Migrate use_motion to use_motion_store and refactor API
Replaces all instances of use_motion with use_motion_store across documentation, examples, and source files. Removes the AnimationManager trait and related helper from src/lib.rs, consolidating animation management into use_motion_store in src/store.rs. Updates documentation and code samples to reflect the new API, improving clarity and consistency for users.
