+++ src/lib.rs
@@ -22,21 +22,21 @@
 //! // Optional: Configure resource pools for optimal performance (recommended for production)
 //! resource_pools::init_high_performance();
 //!
-//! let mut value = use_motion(0.0f32);
+//! let mut motion = use_motion_store(0.0f32);
 //!
 //! // Basic animation - automatically uses all optimizations
-//! value.animate_to(100.0, AnimationConfig::new(AnimationMode::Spring(Spring::default())));
+//! motion.animate_to(100.0, AnimationConfig::new(AnimationMode::Spring(Spring::default())));
 //!
 //! // Animation with custom epsilon for fine-tuned performance (optional)
-//! value.animate_to(
+//! motion.animate_to(
 //!     100.0,
 //!     AnimationConfig::new(AnimationMode::Spring(Spring::default()))
 //!         .with_epsilon(0.001) // Tighter threshold for high-precision animations
 //! );
 //!
 //! // Check if animation is running
-//! if value.is_running() {
-//!     println!("Animation is active with current value: {}", value.get_value());
+//! if motion.is_running() {
+//!     println!("Animation is active with current value: {}", motion.get_value());
 //! }
 //! ```
 //!
@@ -98,12 +98,10 @@
 #![deny(clippy::option_if_let_else)] // Prefer map/and_then
 
 use animations::core::Animatable;
-use dioxus::prelude::*;
 pub use instant::Duration;
 
 pub mod animations;
 pub mod keyframes;
-pub mod manager;
 pub mod motion;
 pub mod pool;
 pub mod sequence;
@@ -117,7 +115,6 @@ pub use dioxus_motion_transitions_macro;
 pub use animations::platform::{MotionTime, TimeProvider};
 
 pub use keyframes::{Keyframe, KeyframeAnimation};
-pub use manager::AnimationManager;
 
 use motion::Motion;
 
@@ -138,7 +135,7 @@ pub mod prelude {
         AnimatableRoute, AnimatedOutlet, AnimatedRouterContext, TransitionVariantResolver,
         use_animated_router,
     };
-    pub use crate::{AnimationManager, Duration, Time, TimeProvider, use_motion};
+    pub use crate::{Duration, Time, TimeProvider};
 
     // Performance optimization exports
     pub use crate::motion::MotionOptimizationStats;
@@ -147,109 +144,3 @@ pub mod prelude {
 }
 
 pub type Time = MotionTime;
-
-/// Helper function to calculate the appropriate delay for the animation loop
-fn calculate_delay(dt: f32, running_frames: u32) -> Duration {
-    #[cfg(feature = "web")]
-    {
-        // running_frames is not used in web builds but kept for API consistency
-        let _ = running_frames;
-        match dt {
-            x if x < 0.008 => Duration::from_millis(8),  // ~120fps
-            x if x < 0.016 => Duration::from_millis(16), // ~60fps
-            _ => Duration::from_millis(32),              // ~30fps
-        }
-    }
-    #[cfg(not(feature = "web"))]
-    {
-        if running_frames <= 200 {
-            Duration::from_micros(8333) // ~120fps
-        } else {
-            match dt {
-                x if x < 0.005 => Duration::from_millis(8),  // ~120fps
-                x if x < 0.011 => Duration::from_millis(16), // ~60fps
-                _ => Duration::from_millis(33),              // ~30fps
-            }
-        }
-    }
-}
-
-/// Creates an animation manager that continuously updates a motion state.
-///
-/// This function initializes a motion state with the provided initial value and spawns an asynchronous loop
-/// that updates the animation state based on the elapsed time between frames. When the animation is running,
-/// it updates the state using the calculated time delta and dynamically adjusts the update interval to optimize CPU usage;
-/// when the animation is inactive, it waits longer before polling again.
-///
-/// # Example
-///
-/// ```no_run
-/// use dioxus_motion::prelude::*;
-/// use dioxus::prelude::*;
-///
-/// fn app() -> Element {
-///     let mut value = use_motion(0.0f32);
-///
-///     // Animate to 100 with spring physics
-///     value.animate_to(
-///         100.0,
-///         AnimationConfig::new(AnimationMode::Spring(Spring::default()))
-///     );
-///
-///     rsx! {
-///         div {
-///             style: "transform: translateY({value.get_value()}px)",
-///             "Animated content"
-///         }
-///     }
-/// }
-/// ```
-pub fn use_motion<T: Animatable + Send + 'static>(initial: T) -> impl AnimationManager<T> {
-    let mut state = use_signal(|| Motion::new(initial));
-
-    #[cfg(feature = "web")]
-    let idle_poll_rate = Duration::from_millis(100);
-
-    #[cfg(not(feature = "web"))]
-    let idle_poll_rate = Duration::from_millis(33);
-
-    use_effect(move || {
-        // This executes after rendering is complete
-        spawn(async move {
-            let mut last_frame = Time::now();
-            let mut running_frames = 0u32;
-
-            loop {
-                let now = Time::now();
-                let dt = (now.duration_since(last_frame).as_secs_f32()).min(0.1);
-                last_frame = now;
-
-                // Only check if running first, then write to the signal
-                if (*state.peek()).is_running() {
-                    running_frames += 1;
-                    let prev_value = (*state.peek()).get_value();
-                    let updated = (*state.write()).update(dt);
-                    let new_value = (*state.peek()).get_value();
-                    let epsilon = (*state.peek()).get_epsilon();
-                    // Only trigger a re-render if the value changed significantly
-                    if (new_value - prev_value).magnitude() > epsilon || updated {
-                        // State has changed enough, continue
-                    } else {
-                        // Skip this frame's update to avoid unnecessary re-render
-                        let delay = calculate_delay(dt, running_frames);
-                        Time::delay(delay).await;
-                        continue;
-                    }
-
-                    let delay = calculate_delay(dt, running_frames);
-                    Time::delay(delay).await;
-                } else {
-                    running_frames = 0;
-                    Time::delay(idle_poll_rate).await;
-                }
-            }
-        });
-    });
-
-    state
-}
